refactor: move tests to component-specific folders
