+++ src/components/Movies/MoviesList.tsx
@@ -1,4 +1,4 @@
-import type { Movie } from '../types/movies/Movie';
+import type { Movie } from '../../types/movies/Movie';
 import MovieCard from './MovieCard';
 
 type MoviesListProps = {
